Homework1/requirements.txt
numpy==1.22.3
Pillow==9.1.0
scikit_image==0.19.2
skimage==0.0
